+++ app/src/test/java/com/inky/fitnesscalendar/ui/views/EditActivityViewTest.kt
@@ -11,17 +11,20 @@ import androidx.test.core.app.ApplicationProvider
 import com.inky.fitnesscalendar.MainApp
 import com.inky.fitnesscalendar.data.ActivityCategory
 import com.inky.fitnesscalendar.data.ContentColor
-import com.inky.fitnesscalendar.db.entities.Activity
+import com.inky.fitnesscalendar.data.EpochDay
+import com.inky.fitnes`scalendar.db.entities.Activity
 import com.inky.fitnesscalendar.db.entities.ActivityType
 import com.inky.fitnesscalendar.db.entities.RichActivity
 import com.inky.fitnesscalendar.localization.LocalizationRepository
 import com.inky.fitnesscalendar.util.toDate
+import com.inky.fitnesscalendar.util.toLocalDateTime
 import org.junit.Assert.*
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
+import java.time.LocalDate
 import java.time.LocalDateTime
 
 @Config(sdk = [Build.VERSION_CODES.UPSIDE_DOWN_CAKE])
@@ -78,5 +81,64 @@ class EditActivityViewTest {
         // The input should now be visible, because an input has been made
         composeTestRule.onNodeWithTag("button-confirm").assertIsEnabled().performClick()
         assertTrue("The save button should have been clicked", success)
+    }
+
+    /**
+     * Tests that creating an activity with an initial day set works correctly
+     */
+    @Test
+    fun test_initial_day_works() {
+        var hasBeenClicked = false
+
+        val context = ApplicationProvider.getApplicationContext<MainApp>()
+
+        val start = LocalDateTime.of(2024, 11, 1, 23, 30)
+        val end = start.plusHours(1)
+
+        val richActivity = RichActivity(
+            activity = Activity(
+                typeId = 1,
+                startTime = start.toDate(),
+                endTime = end.toDate()
+            ),
+            type = ActivityType(
+                uid = 1,
+                hasDuration = false,
+                activityCategory = ActivityCategory.Sports,
+                name = "TestActivityType",
+                emoji = ":)",
+                color = ContentColor.Color1
+            ),
+            place = null,
+        )
+        composeTestRule.setContent {
+            NewActivity(
+                richActivity = richActivity,
+                localizationRepository = LocalizationRepository(context),
+                onSave = {
+                    assertEquals(
+                        LocalDateTime.of(2024, 12, 31, 23, 30),
+                        it.activity.startTime.toLocalDateTime(),
+                    )
+                    assertEquals(
+                        LocalDateTime.of(2025, 1, 1, 0, 30),
+                        it.activity.endTime.toLocalDateTime()
+                    )
+                    hasBeenClicked = true
+                },
+                onNavigateBack = {},
+                onNavigateNewPlace = {},
+                isTest = true,
+                initialDay = EpochDay(day = LocalDate.of(2024, 12, 31).toEpochDay())
+            )
+        }
+
+        // Initially, the confirm button should not exist, because no inputs have been made
+        composeTestRule.onNodeWithTag("button-confirm").assertDoesNotExist()
+        composeTestRule.onNodeWithTag("input-description").performTextInput("Foo")
+
+        // The input should now be visible, because an input has been made
+        composeTestRule.onNodeWithTag("button-confirm").assertIsEnabled().performClick()
+        assertTrue("The save button should have been clicked", hasBeenClicked)
     }
 }