+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - Fix bug where saving an activity was not possible in some cases if the activity type had no duration
+- Creating a new activity from the day view now defaults it to the correct day
 
 # 2024.10.3